Fixed narrowing conversion

+++ src/NodeSystem/NodeOutputInt.cpp
@@ -9,7 +9,7 @@ NodeOutputInt::NodeOutputInt(Node* parent) : NodeOutput(parent)
 
 NodeDataInt* NodeOutputInt::getData()
 {
-    int index{std::distance(parent->outputs.begin(), std::find(parent->outputs.begin(), parent->outputs.end(), this))};
+    long int index{std::distance(parent->outputs.begin(), std::find(parent->outputs.begin(), parent->outputs.end(), this))};
     NodeData* rawData{parent->relations[index](&(parent->inputs))};
     NodeDataInt* data{dynamic_cast<NodeDataInt*>(rawData)};
     if(data != nullptr)
